So it begins
Welcome to the Warzone™

ClassAdam implements the Adam optimisation algorithm with a terrible, horrible, no good at all hotrodded decay rate introduced by yours truly.

OptToF is the main addition that certainly does something. I * think * it runs adam on a starting point with a given gradient function (currently hardcoded), but I'm not quite sure that, or why it works. All praise the omnissiah.

Experiments is my experiment file that simulates end user usage. Its currently full of data I stole from the Tutorial. When run, it runs OptToF with these settings and also sometimes works.
Note: Because of the way AlgoToF has print statements, tqdm will spam your console (even if the prints are skipped). You have been warned

FunctionGenerators is an experiment file that looks at different ways of generating starting functions.

The only file of original ToF that is modified is ClassToF to expose the user to run_dens_opt()

Note: the core functionality of all this here was programmed in approximately 0.9 days, therefore it must be good.

TODO: None of my choices are motivated, the code is undocumented and messy, nothing is properly implemented yet, none of the options are exposed to the user (including variables, choosing the optimiser, choosing the parameter variant), the optimiser cant intelligently stop and you cant queue multiple runs, the starting point is idiotic, its not running in parallel which it absolutely could be, and I'm not even sure the gradient is pointing downwards. Yay :D